# iOS-3065 Improve code block performance