First block:
before: `~235 ms`
after: `~ 35 ms`

Folowing blocks:
before: `~235 ms`
after: `~ 3 ms`


Aslo I deleted double-update for the all blocks on init ✨